Taller_6/requirements.txt
"docker" 
redis-py 
Flask 
"keydb"